+++ changelog.md
@@ -1,3 +1,7 @@
+20170607
+
+- show total stats
+
 20170606
 
 - adding mesh selection stats (verts & tri)